+++ packages/core-utils/src/index.ts
@@ -2,7 +2,6 @@ export * from './format/number-format';
 export * from './io-utils';
 export * from './logic-utils';
 export * from './nav/nav-utils';
-export * from './promise-utils';
 export * from './state-utils';
 export * from './timer-utils';
 export * from './type-utils/deep-partial';
