+++ src/components/bucket-utils.js
@@ -1,4 +1,3 @@
-import { abandonedPromise } from '@terra-ui-packages/core-utils';
 import _ from 'lodash/fp';
 import { useState } from 'react';
 import { h } from 'react-hyperscript-helpers';
@@ -13,7 +12,7 @@ export const withRequesterPaysHandler = _.curry((handler, fn) => async (...args)
   } catch (error) {
     if (error.requesterPaysError) {
       handler();
-      return abandonedPromise();
+      return Utils.abandonedPromise();
     }
     throw error;
   }
