+++ src/workspace-data/data-table/wds/wds-status.test.ts
@@ -1,7 +1,8 @@
-import { abandonedPromise, DeepPartial } from '@terra-ui-packages/core-utils';
+import { DeepPartial } from '@terra-ui-packages/core-utils';
 import { Ajax } from 'src/libs/ajax';
 import { ListAppResponse } from 'src/libs/ajax/leonardo/models/app-models';
 import { WDSCloneStatusResponse } from 'src/libs/ajax/WorkspaceDataService';
+import { abandonedPromise } from 'src/libs/utils';
 import { asMockedFn, renderHookInAct } from 'src/testing/test-utils';
 
 import { useWdsStatus } from './wds-status';
